refactor(cli): import data-transfer entry points directly
`backup` and `migrate` reached `runExportCommand` / `runImportCommand` through
the `../data-transfer` barrel. Pre-existing (both trace to `786b5735d`); they
surfaced now only because this branch touched those files and the review bot
reads changed hunks.

Kept out of the feature commits, per the house rule that pre-existing lint debt
travels separately.

Note for whoever reads the bot comment: its stated rationale — "ships extra code
to your users & slows page load" — is a browser-bundle concern and does not
apply to a Node CLI handler. Direct imports are still the clearer spelling, so
this follows the advice without adopting the reasoning.

Four other call sites still use the barrel (`export`, `import`, `seed`); left
alone because nothing flagged them and widening this commit would bury the two
that were.
