+++ packages/cli/src/commands/backup/handler.ts
@@ -36,8 +36,9 @@ import { EXIT_CODE } from "../../util/exit-code";
 import type { Logger } from "../../util/logger";
 import { isJsonFormat, loggerForFormat, printJson, validateOutputFormat } from "../../util/output-format";
 import { resolveProductionWorkerUrl } from "../../util/resolve-target";
-import type { StreamingFetchLike } from "../data-transfer";
-import { runExportCommand, runImportCommand } from "../data-transfer";
+import { runExportCommand } from "../data-transfer/export";
+import { runImportCommand } from "../data-transfer/import";
+import type { StreamingFetchLike } from "../data-transfer/shared";
 import type { FetchLike } from "../run/handler";
 import { readAndLogBody } from "../run/handler";
 import type { BackupDestination } from "./destination";
