+++ packages/cli/src/commands/migrate/handler.ts
@@ -45,7 +45,8 @@ import { diffSnapshots, renderMigrationFile } from "../../util/migration-diff";
 import { isJsonFormat, loggerForFormat, printJson, validateOutputFormat } from "../../util/output-format";
 import { resolveProductionWorkerUrl } from "../../util/resolve-target";
 import schemaIrToSnapshot from "../../util/schema-snapshot";
-import { runExportCommand, runImportCommand } from "../data-transfer";
+import { runExportCommand } from "../data-transfer/export";
+import { runImportCommand } from "../data-transfer/import";
 import type { FetchLike } from "../run/handler";
 import { readAndLogBody } from "../run/handler";
 import type { MigrateOptions } from "./index";
